requirements.txt
streamlit
Pillow
opencv-python
scipy